requirements.txt
tweepy==4.14.0
Pillow
imageio
imageio-ffmpeg
numpy
